refactor: rename mod handlers to packet_handler
Reactor in progress. I want to extract the request handlers into a new
mod.
